requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
gunicorn
openpyxl
python-dotenv
